removed cmd UI from MainProgram runner to floorsubsystem and adding main to floorsubsystem
