+++ 3303Elevator/src/app/FloorSubsystem/ScheduledElevatorRequest.java
@@ -6,9 +6,10 @@
  * #ID 101127203 
  */
 package app.FloorSubsystem;
+import java.io.Serializable;
 import java.time.LocalTime;
 import java.time.temporal.ChronoUnit; 
-public class ScheduledElevatorRequest {
+public class ScheduledElevatorRequest implements Serializable {
 	
 	private LocalTime time; //time of request
 	private long millisecondDelay;
